Pattern Hatching
Author John Vlissides is a member of the Gang of Four writers. He has an excellent writing style that is very engaging for this type of material. Besides going into some additional design patterns, and some pattern strengths and pitfalls, you get a fascinating behind the scenes look at what went on in the development stages of the original Design Patterns book.